requirements.txt
openai
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
faiss-cpu
streamlit
python-dotenv